requirements.txt
stormssh
PyQt5
envparse
